Se elimina PersonajesTest y se reestructura la estructura de carpetas
