--- dominio/src/Game/Personaje.java
@@ -1,6 +1,0 @@
-package Game;
-
-public abstract class Personaje
-{
-	protected String nombre;
-}
