+++ dominio/src/main/java/Items/CascoDeMagneto.java
