+++ dominio/src/main/java/Items/ItemEquipable.java
