+++ dominio/src/main/java/Items/ItemMano.java
