+++ dominio/src/main/java/Items/Simbionte.java
