+++ dominio/src/main/java/StatListeners/DanoMagicoListener.java
