+++ dominio/src/main/java/StatListeners/EnergiaMaximaListener.java
